+++ Changelog.md
@@ -33,3 +33,4 @@
 - added Pakettikauppa block in order edit page
 - fixed additional service usage according to delivery method
 - added variables to emails
+- added custom Pakettikauppa order states